requirements.txt
pypdf
langchain
langchain-community
chromadb
pytest
# google-generativeai
google-genai
numpy
streamlit